Issue #2875 Add option to login to redhat.registry.io
